requirements.txt
pymongo[srv]==4.2.0
pandas
numpy
scikit-learn
pymongo
python-dotenv
PyYAML
dill
scipy
imblearn==0.0
blinker
click
colorama
Flask
gunicorn
itsdangerous
Jinja2
Werkzeug
-e . #editable mode